Add files via upload
The introduction glances over a brief view of what we will lean to do in Maya. we stat off with the project of building a space ship.

To navigate the 3D viewport there are several options for doing so. by pressing space bar you can enter a planar view allowing you to view your scene from the x, y, and z co-ordinates as well as standard view. if you press the alt key and right click you gain access to rotation of the camera. left click will zoom the focus and button 3 allows you to move the plane of view without panning the camera.

The interface has many different aspects to it allowing you to create primitive polygons, base shapes to be modeled into the more complex objects, and manipulate them in all sorts of manners. The top bar that holds the options of file, edit, create, ect. is affected by the menu option chosen just below the file option to suit the task you are aiming for. You then also have the tools to transform and manipulate your objects under the toolbox option and scr